requirements.txt
agno
python-dotenv
yfinance
packaging
ddgs
fastapi
uvicorn
google-genai
google-generativeai
streamlit

